Process To, From, CC fields to reflect corrected triplets.
Code correctly transforms triplets in the dataframe to a reduced set of
triplets. Thus, email and name duplications are removed. If a
triplet had an email with no name, the name was obtained from the
reduced set of triplets, and if the triplet had a name with no email,
the email was obtained from d_final, a dictionary keyed on the first,last
names.  If a person had multiple emails, this is not destroyed.
The relation between users and emails is currently 1-1. Different emails
will only appear if the names are mispelled.

Different email accounts associated with a given username are reduced to a
single amail account. This, valuable information is lost. I am not
sure how this happens exactly, BUT FOR NOW, it MAY be ok.

It may be important to manualy change the spelling of a user's name by
adding a suffix to disambiguate between multiple email accounts.
Thus, 'adam','corey' could be transformed to {'adam','corey_m1'} and
{'adam', 'corey_m2'} with a different email account for each. This
scheme would have to be reflected in the d_final dictionary.

Where to implement this scheme is still unknown.